requirements.txt
asgiref==3.6.0
Django==4.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
